# Cache for parsing the schema