@patsissons 2018-10-19
looks like this has broken support for `graphql-config` with multiple schemas defined, as the cache hits are short circuiting schema selection. I'll write up a proper issue and create a PR to fix this shortly.